[#5] mod: 엔티티별 하위 패키지 구분
