+++ com.spring.dozen.delivery/src/main/java/com/spring/dozen/delivery/application/dto/delivery/DeliveryCreate.java
@@ -1,4 +1,4 @@
-package com.spring.dozen.delivery.application.dto;
+package com.spring.dozen.delivery.application.dto.delivery;
 
 public record DeliveryCreate(
 	String orderId,
