+++ com.spring.dozen.delivery/src/main/java/com/spring/dozen/delivery/application/dto/deliveryStaff/CompanyDeliveryStaffCreate.java
@@ -1,4 +1,4 @@
-package com.spring.dozen.delivery.application.dto;
+package com.spring.dozen.delivery.application.dto.deliveryStaff;
 
 public record CompanyDeliveryStaffCreate(
 	Long deliveryStaffId,
