+++ com.spring.dozen.delivery/src/main/java/com/spring/dozen/delivery/application/dto/deliveryStaff/HubDeliveryStaffCreate.java
@@ -1,4 +1,4 @@
-package com.spring.dozen.delivery.application.dto;
+package com.spring.dozen.delivery.application.dto.deliveryStaff;
 
 public record HubDeliveryStaffCreate(
 	Long deliveryStaffId
