+++ com.spring.dozen.delivery/src/main/java/com/spring/dozen/delivery/infra/deliveryStaff/DeliveryStaffHubRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.spring.dozen.delivery.infra;
+package com.spring.dozen.delivery.infra.deliveryStaff;
 
 import java.util.Optional;
 import java.util.UUID;
