+++ com.spring.dozen.delivery/src/main/java/com/spring/dozen/delivery/infra/deliveryStaff/DeliveryStaffRepositoryCustom.java
@@ -1,9 +1,9 @@
-package com.spring.dozen.delivery.infra;
+package com.spring.dozen.delivery.infra.deliveryStaff;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import com.spring.dozen.delivery.presentation.dto.DeliveryStaffSearchCond;
+import com.spring.dozen.delivery.presentation.dto.deliveryStaff.DeliveryStaffSearchCond;
 import com.spring.dozen.delivery.domain.entity.DeliveryStaff;
 
 public interface DeliveryStaffRepositoryCustom {
