+++ com.spring.dozen.delivery/src/main/java/com/spring/dozen/delivery/presentation/dto/deliveryStaff/DeliveryStaffSearchCond.java
@@ -1,4 +1,4 @@
-package com.spring.dozen.delivery.presentation.dto;
+package com.spring.dozen.delivery.presentation.dto.deliveryStaff;
 
 public record DeliveryStaffSearchCond(
 	String staffType,
