+++ com.spring.dozen.delivery/src/main/java/com/spring/dozen/delivery/presentation/dto/deliveryStaff/HubDeliveryStaffCreateRequest.java
@@ -1,0 +1,14 @@
+package com.spring.dozen.delivery.presentation.dto.deliveryStaff;
+
+import com.spring.dozen.delivery.application.dto.deliveryStaff.HubDeliveryStaffCreate;
+
+import jakarta.validation.constraints.NotNull;
+
+public record HubDeliveryStaffCreateRequest(
+	@NotNull
+	Long deliveryStaffId
+) {
+	public HubDeliveryStaffCreate toServiceDto() {
+		return new HubDeliveryStaffCreate(deliveryStaffId);
+	}
+}
